+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Added a count indicator to filterable list tab names
 - Improved trigger conflict detection
+- Moved trigger editor button to left side of field for consistency
 
 ## 2.4.0